font: Implement 'float' scale text rendering (#1040)
'Float' scale text rendering is necessary for the shop popup.

Now ui.Credits uses font.DrawTextLang so that the original
character plane is no longer required.

This change also fixes the algorithm of (*Credits).Draw for
efficient rendering.

This change slightly changes the lineheight of the credits
to use the 'common' lineheight defined at the font package.

+++ internal/ui/label.go
@@ -71,5 +71,5 @@ func (l *Label) DrawAsChild(screen *ebiten.Image, offsetX, offsetY int) {
 
 	tx := (l.x + offsetX) * consts.TileScale
 	ty := (l.y + offsetY) * consts.TileScale
-	font.DrawText(screen, l.Text, tx, ty, int(float64(consts.TextScale)*l.Scale), l.TextAlign, l.Color, len([]rune(l.Text)))
+	font.DrawText(screen, l.Text, tx, ty, float64(consts.TextScale)*l.Scale, l.TextAlign, l.Color, len([]rune(l.Text)))
 }
